refactor: change hooks folder structure
