+++ front/src/components/atoms/DebounedSearch/DebouncedSearch.jsx
@@ -1,7 +1,7 @@
 import React, { useEffect, useState } from 'react';
 import PropTypes from 'prop-types';
 import { Input } from 'antd';
-import useDebounce from '../../../hooks/useDebounce';
+import { useDebounce } from '@sb-ui/hooks/useDebounce';
 
 const { Search } = Input;
