+++ front/src/hooks/useAuthentication.js
@@ -1,7 +1,7 @@
 import { useState } from 'react';
 import { useTranslation } from 'react-i18next';
 import { useHistory } from 'react-router-dom';
-import { setJWT } from '../../utils/jwt';
+import { setJWT } from '../utils/jwt';
 
 const getTranslationFromMessageData = (t, data) => {
   const { key, message } = data;
@@ -12,7 +12,7 @@ const getTranslationFromMessageData = (t, data) => {
   return text;
 };
 
-const useAuthentication = (requestFunc) => {
+export const useAuthentication = (requestFunc) => {
   const { t } = useTranslation();
   const history = useHistory();
   const [error, setError] = useState(null);
@@ -43,5 +43,3 @@ const useAuthentication = (requestFunc) => {
 
   return [auth, error, setError, loading];
 };
-
-export default useAuthentication;
