--- front/src/hooks/useAuthentication/index.js
@@ -1,3 +1,0 @@
-import useAuthentication from './useAuthentication';
-
-export default useAuthentication;
