+++ front/src/hooks/useDebounce.js
@@ -1,6 +1,6 @@
 import { useEffect, useState } from 'react';
 
-const useDebounce = (value, delay) => {
+export const useDebounce = (value, delay) => {
   const [debouncedValue, setDebouncedValue] = useState(value);
 
   useEffect(() => {
@@ -14,5 +14,3 @@ const useDebounce = (value, delay) => {
 
   return debouncedValue;
 };
-
-export default useDebounce;
