--- front/src/hooks/useDebounce/index.js
@@ -1,3 +1,0 @@
-import useDebounce from './useDebounce';
-
-export default useDebounce;
