--- front/src/hooks/useLessonsRequest/index.js
@@ -1,3 +1,0 @@
-import useLessonsRequest from './useLessonsRequest';
-
-export default useLessonsRequest;
