--- front/src/hooks/usePasswordInput/index.js
@@ -1,3 +1,0 @@
-import usePasswordInput from './usePasswordInput';
-
-export default usePasswordInput;
