--- front/src/hooks/useTableRequest/index.js
@@ -1,3 +1,0 @@
-import useTableRequest from './useTableRequest';
-
-export default useTableRequest;
