+++ front/src/pages/SignUp/SignUpForm/SignUpForm.jsx
@@ -2,9 +2,9 @@ import React, { useMemo } from 'react';
 import { Alert, Form, Input } from 'antd';
 import { useTranslation } from 'react-i18next';
 import { useHistory } from 'react-router-dom';
+import { usePasswordInput } from '@sb-ui/hooks/usePasswordInput';
+import { useAuthentication } from '@sb-ui/hooks/useAuthentication';
 import PasswordStrengthIndicator from '../../../components/atoms/PasswordStrengthIndicator';
-import usePasswordInput from '../../../hooks/usePasswordInput';
-import useAuthentication from '../../../hooks/useAuthentication';
 import { postSignUp } from '../../../utils/api/v1/user';
 import { DivAlignCenter, SubmitButton, LinkButton } from './SignUpForm.styled';
 import { SIGN_IN } from '../../../utils/paths';
